Move quant functions to model_quant.py
