+++ modules/model_quant.py
@@ -1,11 +1,17 @@
 import sys
+import copy
+import time
 import diffusers
-from installer import install, log
+from installer import install, log, setup_logging
 
 
-bnb = None
-quanto = None
 ao = None
+bnb = None
+intel_nncf = None
+optimum_quanto = None
+
+quant_last_model_name = None
+quant_last_model_device = None
 
 
 def get_quant(name):
@@ -114,22 +120,42 @@ def load_bnb(msg='', silent=False):
 
 def load_quanto(msg='', silent=False):
     from modules import shared
-    global quanto # pylint: disable=global-statement
-    if quanto is not None:
-        return quanto
+    global optimum_quanto # pylint: disable=global-statement
+    if optimum_quanto is not None:
+        return optimum_quanto
     install('optimum-quanto==0.2.6', quiet=True)
     try:
-        from optimum import quanto as optimum_quanto # pylint: disable=no-name-in-module
-        quanto = optimum_quanto
+        from optimum import quanto # pylint: disable=no-name-in-module
+        optimum_quanto = quanto
         fn = f'{sys._getframe(2).f_code.co_name}:{sys._getframe(1).f_code.co_name}' # pylint: disable=protected-access
         log.debug(f'Quantization: type=quanto version={quanto.__version__} fn={fn}') # pylint: disable=protected-access
         if shared.opts.diffusers_offload_mode in {'balanced', 'sequential'}:
             shared.log.error(f'Quantization: type=quanto offload={shared.opts.diffusers_offload_mode} not supported')
-        return quanto
+        return optimum_quanto
     except Exception as e:
         if len(msg) > 0:
             log.error(f"{msg} failed to import optimum.quanto: {e}")
-        quanto = None
+        optimum_quanto = None
+        if not silent:
+            raise
+    return None
+
+
+def load_nncf(msg='', silent=False):
+    global intel_nncf # pylint: disable=global-statement
+    if intel_nncf is not None:
+        return intel_nncf
+    install('nncf==2.7.0', quiet=True)
+    try:
+        import nncf
+        intel_nncf = nncf
+        fn = f'{sys._getframe(2).f_code.co_name}:{sys._getframe(1).f_code.co_name}' # pylint: disable=protected-access
+        log.debug(f'Quantization: type=nncf version={nncf.__version__} fn={fn}') # pylint: disable=protected-access
+        return intel_nncf
+    except Exception as e:
+        if len(msg) > 0:
+            log.error(f"{msg} failed to import nncf: {e}")
+        intel_nncf = None
         if not silent:
             raise
     return None
@@ -175,3 +201,194 @@ def apply_layerwise(sd_model, quiet:bool=False):
                     log.quiet(quiet, f'Quantization: type=layerwise module={module} cls={cls} storage={storage_dtype} compute={devices.dtype} blocking={not non_blocking}')
         except Exception as e:
             shared.log.error(f'Quantization: type=layerwise {e}')
+
+
+def nncf_send_to_device(model, device):
+    for child in model.children():
+        if child.__class__.__name__ == "WeightsDecompressor":
+            child.scale = child.scale.to(device)
+            child.zero_point = child.zero_point.to(device)
+        nncf_send_to_device(child, device)
+
+
+def nncf_compress_model(model, op=None, sd_model=None):
+    from modules import devices, shared
+    global quant_last_model_name, quant_last_model_device # pylint: disable=global-statement
+    nncf = load_nncf('Quantize model: type=NNCF')
+    model.eval()
+    backup_embeddings = None
+    if hasattr(model, "get_input_embeddings"):
+        backup_embeddings = copy.deepcopy(model.get_input_embeddings())
+    model = nncf.compress_weights(model)
+    nncf_send_to_device(model, devices.device)
+    if hasattr(model, "set_input_embeddings") and backup_embeddings is not None:
+        model.set_input_embeddings(backup_embeddings)
+    if op is not None and shared.opts.quant_shuffle_weights:
+        if quant_last_model_name is not None:
+            if "." in quant_last_model_name:
+                last_model_names = quant_last_model_name.split(".")
+                getattr(getattr(sd_model, last_model_names[0]), last_model_names[1]).to(quant_last_model_device)
+            else:
+                getattr(sd_model, quant_last_model_name).to(quant_last_model_device)
+            devices.torch_gc(force=True)
+        if shared.cmd_opts.medvram or shared.cmd_opts.lowvram or shared.opts.diffusers_offload_mode != "none":
+            quant_last_model_name = op
+            quant_last_model_device = model.device
+        else:
+            quant_last_model_name = None
+            quant_last_model_device = None
+        model.to(devices.device)
+    devices.torch_gc(force=True)
+    return model
+
+
+def nncf_compress_weights(sd_model):
+    try:
+        t0 = time.time()
+        from modules import shared, devices, sd_models
+        shared.log.info(f"Quantization: type=NNCF modules={shared.opts.nncf_compress_weights}")
+        global quant_last_model_name, quant_last_model_device # pylint: disable=global-statement
+
+        sd_model = sd_models.apply_function_to_model(sd_model, nncf_compress_model, shared.opts.nncf_compress_weights, op="nncf")
+        if quant_last_model_name is not None:
+            if "." in quant_last_model_name:
+                last_model_names = quant_last_model_name.split(".")
+                getattr(getattr(sd_model, last_model_names[0]), last_model_names[1]).to(quant_last_model_device)
+            else:
+                getattr(sd_model, quant_last_model_name).to(quant_last_model_device)
+            devices.torch_gc(force=True)
+        quant_last_model_name = None
+        quant_last_model_device = None
+
+        t1 = time.time()
+        shared.log.info(f"Quantization: type=NNCF time={t1-t0:.2f}")
+    except Exception as e:
+        shared.log.warning(f"Quantization: type=NNCF {e}")
+    return sd_model
+
+
+def optimum_quanto_model(model, op=None, sd_model=None, weights=None, activations=None):
+    from modules import devices, shared
+    quanto = load_quanto('Quantize model: type=Optimum Quanto')
+    global quant_last_model_name, quant_last_model_device # pylint: disable=global-statement
+    if sd_model is not None and "Flux" in sd_model.__class__.__name__: # LayerNorm is not supported
+        exclude_list = ["transformer_blocks.*.norm1.norm", "transformer_blocks.*.norm2", "transformer_blocks.*.norm1_context.norm", "transformer_blocks.*.norm2_context", "single_transformer_blocks.*.norm.norm", "norm_out.norm"]
+    else:
+        exclude_list = None
+    weights = getattr(quanto, weights) if weights is not None else getattr(quanto, shared.opts.optimum_quanto_weights_type)
+    if activations is not None:
+        activations = getattr(quanto, activations) if activations != 'none' else None
+    elif shared.opts.optimum_quanto_activations_type != 'none':
+        activations = getattr(quanto, shared.opts.optimum_quanto_activations_type)
+    else:
+        activations = None
+    model.eval()
+    backup_embeddings = None
+    if hasattr(model, "get_input_embeddings"):
+        backup_embeddings = copy.deepcopy(model.get_input_embeddings())
+    quanto.quantize(model, weights=weights, activations=activations, exclude=exclude_list)
+    quanto.freeze(model)
+    if hasattr(model, "set_input_embeddings") and backup_embeddings is not None:
+        model.set_input_embeddings(backup_embeddings)
+    if op is not None and shared.opts.quant_shuffle_weights:
+        if quant_last_model_name is not None:
+            if "." in quant_last_model_name:
+                last_model_names = quant_last_model_name.split(".")
+                getattr(getattr(sd_model, last_model_names[0]), last_model_names[1]).to(quant_last_model_device)
+            else:
+                getattr(sd_model, quant_last_model_name).to(quant_last_model_device)
+            devices.torch_gc(force=True)
+        if shared.cmd_opts.medvram or shared.cmd_opts.lowvram or shared.opts.diffusers_offload_mode != "none":
+            quant_last_model_name = op
+            quant_last_model_device = model.device
+        else:
+            quant_last_model_name = None
+            quant_last_model_device = None
+        model.to(devices.device)
+    devices.torch_gc(force=True)
+    return model
+
+
+def optimum_quanto_weights(sd_model):
+    try:
+        t0 = time.time()
+        from modules import shared, devices, sd_models
+        if shared.opts.diffusers_offload_mode in {"balanced", "sequential"}:
+            shared.log.warning(f"Quantization: type=Optimum.quanto offload={shared.opts.diffusers_offload_mode} not compatible")
+            return sd_model
+        shared.log.info(f"Quantization: type=Optimum.quanto: modules={shared.opts.optimum_quanto_weights}")
+        global quant_last_model_name, quant_last_model_device # pylint: disable=global-statement
+        quanto = load_quanto()
+        quanto.tensor.qbits.QBitsTensor.create = lambda *args, **kwargs: quanto.tensor.qbits.QBitsTensor(*args, **kwargs)
+
+        sd_model = sd_models.apply_function_to_model(sd_model, optimum_quanto_model, shared.opts.optimum_quanto_weights, op="optimum-quanto")
+        if quant_last_model_name is not None:
+            if "." in quant_last_model_name:
+                last_model_names = quant_last_model_name.split(".")
+                getattr(getattr(sd_model, last_model_names[0]), last_model_names[1]).to(quant_last_model_device)
+            else:
+                getattr(sd_model, quant_last_model_name).to(quant_last_model_device)
+            devices.torch_gc(force=True)
+        quant_last_model_name = None
+        quant_last_model_device = None
+
+        if shared.opts.optimum_quanto_activations_type != 'none':
+            activations = getattr(quanto, shared.opts.optimum_quanto_activations_type)
+        else:
+            activations = None
+
+        if activations is not None:
+            def optimum_quanto_freeze(model, op=None, sd_model=None): # pylint: disable=unused-argument
+                quanto.freeze(model)
+                return model
+            if shared.opts.diffusers_offload_mode == "model":
+                sd_model.enable_model_cpu_offload(device=devices.device)
+                if hasattr(sd_model, "encode_prompt"):
+                    original_encode_prompt = sd_model.encode_prompt
+                    def encode_prompt(*args, **kwargs):
+                        embeds = original_encode_prompt(*args, **kwargs)
+                        sd_model.maybe_free_model_hooks() # Diffusers keeps the TE on VRAM
+                        return embeds
+                    sd_model.encode_prompt = encode_prompt
+            else:
+                sd_models.move_model(sd_model, devices.device)
+            with quanto.Calibration(momentum=0.9):
+                sd_model(prompt="dummy prompt", num_inference_steps=10)
+            sd_model = sd_models.apply_function_to_model(sd_model, optimum_quanto_freeze, shared.opts.optimum_quanto_weights, op="optimum-quanto-freeze")
+            if shared.opts.diffusers_offload_mode == "model":
+                sd_models.disable_offload(sd_model)
+                sd_models.move_model(sd_model, devices.cpu)
+                if hasattr(sd_model, "encode_prompt"):
+                    sd_model.encode_prompt = original_encode_prompt
+            devices.torch_gc(force=True)
+
+        t1 = time.time()
+        shared.log.info(f"Quantization: type=Optimum.quanto time={t1-t0:.2f}")
+    except Exception as e:
+        shared.log.warning(f"Quantization: type=Optimum.quanto {e}")
+    return sd_model
+
+
+def torchao_quantization(sd_model):
+    from modules import shared, devices, sd_models
+    torchao = load_torchao()
+    q = torchao.quantization
+
+    fn = getattr(q, shared.opts.torchao_quantization_type, None)
+    if fn is None:
+        shared.log.error(f"Quantization: type=TorchAO type={shared.opts.torchao_quantization_type} not supported")
+        return sd_model
+    def torchao_model(model, op=None, sd_model=None): # pylint: disable=unused-argument
+        q.quantize_(model, fn(), device=devices.device)
+        return model
+
+    shared.log.info(f"Quantization: type=TorchAO pipe={sd_model.__class__.__name__} quant={shared.opts.torchao_quantization_type} fn={fn} targets={shared.opts.torchao_quantization}")
+    try:
+        t0 = time.time()
+        sd_models.apply_function_to_model(sd_model, torchao_model, shared.opts.torchao_quantization, op="torchao")
+        t1 = time.time()
+        shared.log.info(f"Quantization: type=TorchAO time={t1-t0:.2f}")
+    except Exception as e:
+        shared.log.error(f"Quantization: type=TorchAO {e}")
+    setup_logging() # torchao uses dynamo which messes with logging so reset is needed
+    return sd_model
