+++ modules/sd_models_utils.py
@@ -1,4 +1,5 @@
 import io
+import copy
 import json
 import inspect
 import os.path
@@ -6,7 +7,7 @@ from rich import progress # pylint: disable=redefined-builtin
 import torch
 import safetensors.torch
 
-from modules import paths, shared, errors
+from modules import paths, shared, devices, errors
 from modules.sd_checkpoint import CheckpointInfo, select_checkpoint, list_models, checkpoints_list, checkpoint_titles, get_closet_checkpoint_match, model_hash, update_model_hashes, setup_model, write_metadata, read_metadata_from_safetensors # pylint: disable=unused-import
 from modules.sd_offload import disable_offload, set_diffuser_offload, apply_balanced_offload, set_accelerate # pylint: disable=unused-import
 from modules.sd_models_legacy import get_checkpoint_state_dict, load_model_weights, load_model, repair_config # pylint: disable=unused-import
@@ -148,4 +149,84 @@ def patch_diffuser_config(sd_model, model_file):
                         component.config.__frozen = False # pylint: disable=protected-access
                     component.config[k] = v
                     updated[k] = v
+    return sd_model
+
+
+def apply_function_to_model(sd_model, function, options, op=None):
+    if "Model" in options or "Transformer" in options:
+        if hasattr(sd_model, 'transformer') and hasattr(sd_model.transformer, 'config'):
+            sd_model.transformer = function(sd_model.transformer, op="transformer", sd_model=sd_model)
+    if "Model" in options:
+        if hasattr(sd_model, 'unet') and hasattr(sd_model.unet, 'config'):
+            sd_model.unet = function(sd_model.unet, op="unet", sd_model=sd_model)
+        if hasattr(sd_model, 'decoder_pipe') and hasattr(sd_model, 'decoder'):
+            sd_model.decoder = None
+            sd_model.decoder = sd_model.decoder_pipe.decoder = function(sd_model.decoder_pipe.decoder, op="decoder_pipe.decoder", sd_model=sd_model)
+        if hasattr(sd_model, 'prior_pipe') and hasattr(sd_model.prior_pipe, 'prior'):
+            if op == "nncf" and "StableCascade" in sd_model.__class__.__name__: # fixes dtype errors
+                backup_clip_txt_pooled_mapper = copy.deepcopy(sd_model.prior_pipe.prior.clip_txt_pooled_mapper)
+            sd_model.prior_pipe.prior = function(sd_model.prior_pipe.prior, op="prior_pipe.prior", sd_model=sd_model)
+            if op == "nncf" and "StableCascade" in sd_model.__class__.__name__:
+                sd_model.prior_pipe.prior.clip_txt_pooled_mapper = backup_clip_txt_pooled_mapper
+    if "Text Encoder" in options:
+        if hasattr(sd_model, 'text_encoder') and hasattr(sd_model.text_encoder, 'config'):
+            if hasattr(sd_model, 'decoder_pipe') and hasattr(sd_model.decoder_pipe, 'text_encoder') and hasattr(sd_model.decoder_pipe.text_encoder, 'config'):
+                sd_model.decoder_pipe.text_encoder = function(sd_model.decoder_pipe.text_encoder, op="decoder_pipe.text_encoder", sd_model=sd_model)
+            else:
+                if op == "nncf" and sd_model.text_encoder.__class__.__name__ in {"T5EncoderModel", "UMT5EncoderModel"}:
+                    from modules.sd_hijack import NNCF_T5DenseGatedActDense # T5DenseGatedActDense uses fp32
+                    for i in range(len(sd_model.text_encoder.encoder.block)):
+                        sd_model.text_encoder.encoder.block[i].layer[1].DenseReluDense = NNCF_T5DenseGatedActDense(
+                            sd_model.text_encoder.encoder.block[i].layer[1].DenseReluDense,
+                            dtype=torch.float32 if devices.dtype != torch.bfloat16 else torch.bfloat16
+                        )
+                sd_model.text_encoder = function(sd_model.text_encoder, op="text_encoder", sd_model=sd_model)
+        if hasattr(sd_model, 'text_encoder_2') and hasattr(sd_model.text_encoder_2, 'config'):
+            if op == "nncf" and sd_model.text_encoder_2.__class__.__name__ in {"T5EncoderModel", "UMT5EncoderModel"}:
+                from modules.sd_hijack import NNCF_T5DenseGatedActDense # T5DenseGatedActDense uses fp32
+                for i in range(len(sd_model.text_encoder_2.encoder.block)):
+                    sd_model.text_encoder_2.encoder.block[i].layer[1].DenseReluDense = NNCF_T5DenseGatedActDense(
+                        sd_model.text_encoder_2.encoder.block[i].layer[1].DenseReluDense,
+                        dtype=torch.float32 if devices.dtype != torch.bfloat16 else torch.bfloat16
+                    )
+            sd_model.text_encoder_2 = function(sd_model.text_encoder_2, op="text_encoder_2", sd_model=sd_model)
+        if hasattr(sd_model, 'text_encoder_3') and hasattr(sd_model.text_encoder_3, 'config'):
+            if op == "nncf" and sd_model.text_encoder_3.__class__.__name__ in {"T5EncoderModel", "UMT5EncoderModel"}:
+                from modules.sd_hijack import NNCF_T5DenseGatedActDense # T5DenseGatedActDense uses fp32
+                for i in range(len(sd_model.text_encoder_3.encoder.block)):
+                    sd_model.text_encoder_3.encoder.block[i].layer[1].DenseReluDense = NNCF_T5DenseGatedActDense(
+                        sd_model.text_encoder_3.encoder.block[i].layer[1].DenseReluDense,
+                        dtype=torch.float32 if devices.dtype != torch.bfloat16 else torch.bfloat16
+                    )
+            sd_model.text_encoder_3 = function(sd_model.text_encoder_3, op="text_encoder_3", sd_model=sd_model)
+        if hasattr(sd_model, 'prior_pipe') and hasattr(sd_model.prior_pipe, 'text_encoder') and hasattr(sd_model.prior_pipe.text_encoder, 'config'):
+            sd_model.prior_pipe.text_encoder = function(sd_model.prior_pipe.text_encoder, op="prior_pipe.text_encoder", sd_model=sd_model)
+    if "VAE" in options:
+        if hasattr(sd_model, 'vae') and hasattr(sd_model.vae, 'decode'):
+            if op == "compile":
+                sd_model.vae.decode = function(sd_model.vae.decode, op="vae_decode", sd_model=sd_model)
+                sd_model.vae.encode = function(sd_model.vae.encode, op="vae_encode", sd_model=sd_model)
+            else:
+                sd_model.vae = function(sd_model.vae, op="vae", sd_model=sd_model)
+        if hasattr(sd_model, 'movq') and hasattr(sd_model.movq, 'decode'):
+            if op == "compile":
+                sd_model.movq.decode = function(sd_model.movq.decode, op="movq_decode", sd_model=sd_model)
+                sd_model.movq.encode = function(sd_model.movq.encode, op="movq_encode", sd_model=sd_model)
+            else:
+                sd_model.movq = function(sd_model.movq, op="movq", sd_model=sd_model)
+        if hasattr(sd_model, 'vqgan') and hasattr(sd_model.vqgan, 'decode'):
+            if op == "compile":
+                sd_model.vqgan.decode = function(sd_model.vqgan.decode, op="vqgan_decode", sd_model=sd_model)
+                sd_model.vqgan.encode = function(sd_model.vqgan.encode, op="vqgan_encode", sd_model=sd_model)
+            else:
+                sd_model.vqgan = function(sd_model.vqgan, op="vqgan", sd_model=sd_model)
+            if hasattr(sd_model, 'decoder_pipe') and hasattr(sd_model.decoder_pipe, 'vqgan'):
+                if op == "compile":
+                    sd_model.decoder_pipe.vqgan.decode = function(sd_model.decoder_pipe.vqgan.decode, op="vqgan_decode", sd_model=sd_model)
+                    sd_model.decoder_pipe.vqgan.encode = function(sd_model.decoder_pipe.vqgan.encode, op="vqgan_encode", sd_model=sd_model)
+                else:
+                    sd_model.decoder_pipe.vqgan = sd_model.vqgan
+        if hasattr(sd_model, 'image_encoder') and hasattr(sd_model.image_encoder, 'config'):
+            sd_model.image_encoder = function(sd_model.image_encoder, op="image_encoder", sd_model=sd_model)
+
     return sd_model
